pyproject.toml
[tool.poetry]
name = "duffelbag"
version = "0.1.0"
description = ""
authors = ["Chromosomologist <chromosomologist@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.0"
arkprts = {git = "https://github.com/thesadru/arkprts"}
piccolo = {git = "https://github.com/piccolo-orm/piccolo", extras = ["postgres"], tag = "1.0a2"}
argon2-cffi = "^21.3.0"
coloredlogs = "^15.0.1"
pillow = "^10.0.1"


[tool.poetry.group.dev.dependencies]
devtools = "^0.11.0"
ruff = "^0.0.289"
pre-commit = "^3.3.2"
pyright = "^1.1.310"
slotscheck = "^0.17.0"


[tool.poetry.group.discord.dependencies]
disnake = "^2.8.1"
disnake-ext-plugins = {git = "https://github.com/DisnakeCommunity/disnake-ext-plugins"}
disnake-ext-components = {git = "https://github.com/DisnakeCommunity/disnake-ext-components", rev = "rewrite"}


[tool.poetry.group.speedups.dependencies]
uvloop = "^0.17.0"
orjson = "^3.8.14"


[tool.poetry.scripts]
"repopulate" = "scripts.repopulate:_sync_main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pyright]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.*",
    "database/piccolo_migrations",
    ".venv",
]
reportMissingTypeStubs=false


[tool.black]
line-length = 100
extend-exclude = ".*database/piccolo_migrations.*"


[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = [
    "database/piccolo_migrations",
]

select = ["ALL"]
ignore = [
    # `self` and `cls` do not need annotations.
    "ANN101", "ANN102",
    # Magic methods do not need documentation.
    "D105",
    # Class docstrings are at the class level.
    "D107",
    # We use assertions as an in-place alternative to `typing.cast`.
    "S101",
    # Pyright error codes are obnoxiously long.
    "PGH003",
    # TODO / HACK comments do not need to be flagged.
    "FIX002", "FIX004",
    # TODO comments do not need to have an author or issue specified.
    "TD002", "TD003",
]


[tool.ruff.per-file-ignores]
# T201: Allow printing in scripts.
"scripts/*" = ["T201"]
# Wildcard imports are fine in __init__; more convenient than duplicating exports and maintaining them.
"__init__.py" = ["F403", "F405"]
# Code lying in wait for composite unique constraints.
"database/populate.py" = ["ERA"]
# `id` is a very common field name in models
"raw_data/models/*" = ["A003"]
"database/models/*" = ["A003"]
